Move block downloader
